Capture the Flag/requirements.txt
pillow
numpy
matplotlib
pytesseract